test(api): moving api tests with fixture into the src directory
