+++ lib/__tests__/general.spec.js
@@ -1,6 +1,6 @@
 // eslint-disable-next-line import/no-unresolved, import/no-extraneous-dependencies
-const {mockDataObject} = require('mock-data'); // Note mock-data live under __mocks__ dir
-const Translate = require('../../lib/translate-json-object');
+const Translate = require('../translate-json-object');
+const mock = require('../fixture/data.fixture');
 
 const TJO = new Translate();
 const init = jest.fn(TJO.init);
@@ -33,7 +33,7 @@ describe('Translate Module TJO: ', () => {
 
 	it('should fail to translate() using unkown service', () => {
 		init({unkownApiService: 'apiToken'});
-		return translate(mockDataObject, 'ar')
+		return translate(mock.dataObject, 'ar')
 						.catch(err => {
 							expect(err).toBeDefined();
 						});
@@ -41,7 +41,7 @@ describe('Translate Module TJO: ', () => {
 
 	it('Should fail to translate because of a missing language param: fail', () => {
 		init({yandexApiKey: 'apiToken'});
-		return translate(mockDataObject)
+		return translate(mock.dataObject)
 						.catch(err => {
 							expect(err).toBeDefined();
 						});
@@ -49,7 +49,7 @@ describe('Translate Module TJO: ', () => {
 
 	it('should TJO.translate fails, becuase of missing google/yandex token', () => {
 		init();
-		return translate(mockDataObject, 'ar')
+		return translate(mock.dataObject, 'ar')
 						.catch(err => {
 							expect(err).toBeDefined();
 						});
